+++ src/main/java/com/example/server/service/RoleService.java
@@ -7,15 +7,15 @@ import java.util.List;
 
 public interface RoleService {
 
-    RoleDto getRoleById(int id);
+    RoleDto getRoleById(Long id);
 
     List<RoleDto> getAllRoles();
 
-    List<Role> getRoleListByIds(List<Integer> ids);
+    List<Role> getRoleListByIds(List<Long> ids);
 
     void addRole(RoleDto role);
 
     void updateRole(RoleDto roleDto);
 
-    void removeRole(int id);
+    void removeRole(Long id);
 }
